+++ kafka-streams/kafka-streams/src/main/scala/com/twitter/finatra/kafkastreams/KafkaStreamsTwitterServer.scala
@@ -60,7 +60,7 @@ abstract class KafkaStreamsTwitterServer
     helpPrefix = "A finagle destination or"
   )
 
-  // Configs using kafka default
+  // StreamsConfig default flags
   private val numStreamThreads =
     flagWithKafkaDefault[Integer](StreamsConfig.NUM_STREAM_THREADS_CONFIG)
   private val numStandbyReplicas =
@@ -71,13 +71,29 @@ abstract class KafkaStreamsTwitterServer
     flagWithKafkaDefault[Long](StreamsConfig.CACHE_MAX_BYTES_BUFFERING_CONFIG)
   private val metadataMaxAge = flagWithKafkaDefault[Long](StreamsConfig.METADATA_MAX_AGE_CONFIG)
 
-  private val consumerMaxPollRecords = consumerFlagWithKafkaDefault[Int](ConsumerConfig.MAX_POLL_RECORDS_CONFIG)
-  private val consumerAutoOffsetReset = consumerFlagWithKafkaDefault[String](ConsumerConfig.AUTO_OFFSET_RESET_CONFIG)
-  private val consumerSessionTimeout = consumerFlagWithKafkaDefault[Int](ConsumerConfig.SESSION_TIMEOUT_MS_CONFIG)
-  private val consumerHeartbeatInterval = consumerFlagWithKafkaDefault[Int](ConsumerConfig.HEARTBEAT_INTERVAL_MS_CONFIG)
-  private val consumerFetchMin = consumerFlagWithKafkaDefault[Int](ConsumerConfig.FETCH_MIN_BYTES_CONFIG)
-  private val consumerFetchMaxWait = consumerFlagWithKafkaDefault[Int](ConsumerConfig.FETCH_MAX_WAIT_MS_CONFIG)
+  // ConsumerConfig default flags
+  private val consumerMaxPollRecords =
+    consumerFlagWithKafkaDefault[Int](ConsumerConfig.MAX_POLL_RECORDS_CONFIG)
+  private val consumerMaxPollInterval =
+    consumerFlagWithKafkaDefault[Int](ConsumerConfig.MAX_POLL_INTERVAL_MS_CONFIG)
+  private val consumerAutoOffsetReset =
+    consumerFlagWithKafkaDefault[String](ConsumerConfig.AUTO_OFFSET_RESET_CONFIG)
+  private val consumerSessionTimeout =
+    consumerFlagWithKafkaDefault[Int](ConsumerConfig.SESSION_TIMEOUT_MS_CONFIG)
+  private val consumerHeartbeatInterval =
+    consumerFlagWithKafkaDefault[Int](ConsumerConfig.HEARTBEAT_INTERVAL_MS_CONFIG)
+  private val consumerFetchMin =
+    consumerFlagWithKafkaDefault[Int](ConsumerConfig.FETCH_MIN_BYTES_CONFIG)
+  private val consumerFetchMaxWait =
+    consumerFlagWithKafkaDefault[Int](ConsumerConfig.FETCH_MAX_WAIT_MS_CONFIG)
+  private val consumerMaxPartitionFetch =
+    consumerFlagWithKafkaDefault[Int](ConsumerConfig.MAX_PARTITION_FETCH_BYTES_CONFIG)
+  private val consumerRequestTimeout =
+    consumerFlagWithKafkaDefault[Int](ConsumerConfig.REQUEST_TIMEOUT_MS_CONFIG)
+  private val consumerConnectionsMaxIdle =
+    consumerFlagWithKafkaDefault[Long](ConsumerConfig.CONNECTIONS_MAX_IDLE_MS_CONFIG)
 
+  // ProducerConfig default flag
   private val producerLinger = producerFlagWithKafkaDefault[Long](ProducerConfig.LINGER_MS_CONFIG)
 
   // Configs with customized default
@@ -241,6 +257,10 @@ abstract class KafkaStreamsTwitterServer
         .consumer.metricsSampleWindow(60.seconds)
         .consumer.autoOffsetReset(OffsetResetStrategy.valueOf(consumerAutoOffsetReset().toUpperCase))
         .consumer.maxPollRecords(consumerMaxPollRecords())
+        .consumer.maxPollInterval(consumerMaxPollInterval().milliseconds)
+        .consumer.maxPartitionFetch(consumerMaxPartitionFetch().bytes)
+        .consumer.requestTimeout(consumerRequestTimeout().milliseconds)
+        .consumer.connectionsMaxIdle(consumerConnectionsMaxIdle().milliseconds)
         .consumer.interceptor[KafkaStreamsMonitoringConsumerInterceptor]
 
     if (applicationId().nonEmpty) {